Android 테스트 코드로 변경 getSystemService를 사용하기 위해서 Android Test로 변경함.
